requirements.txt
accelerate==1.3.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
appnope==0.1.4
argcomplete==3.6.0
asttokens==3.0.0
asyncio==3.4.3
attrs==25.3.0
audioread==3.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.37.14
botocore==1.37.14
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chronos-forecasting
click==8.1.8
cohere==5.14.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.8
fastavro==1.10.0
fastmcp==2.3.3
ffmpeg-python==0.2.0
filelock==3.18.0
fonttools==4.55.8
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.1
graphviz==0.20.3
griffe==1.6.0
groq==0.19.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpath-python==1.0.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
lockfile==0.12.2
logfire==3.8.1
logfire-api==3.9.0
luigi==3.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mcp==1.9.0
mcp-client==0.2.0
mdurl==0.1.2
mistralai==1.5.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
mypy-extensions==1.0.0
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
numba==0.61.0
numpy==2.2.3
openai==1.79.0
openai-agents==0.0.15
openapi-pydantic==0.5.1
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playaudio==1.0.5
pooch==1.8.2
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic-ai==0.0.41
pydantic-ai-slim==0.0.41
pydantic-graph==0.0.41
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
pypdfium2==4.30.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.2
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.2
typer==0.15.3
types-requests==2.32.0.20250306
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
whisper==1.1.10
whispercpp==0.0.17
wrapt==1.17.2
xmltodict==0.14.2
zipp==3.21.0
